# swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Admin"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/task": {
      "delete": {
        "operationId": "Admin_TaskDelete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTaskDeleteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiTaskDeleteRequest"
            }
          }
        ],
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "operationId": "Admin_TaskCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTaskCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiTaskCreateRequest"
            }
          }
        ],
        "tags": [
          "Admin"
        ]
      },
      "put": {
        "operationId": "Admin_TaskUpdate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTaskUpdateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiTaskUpdateRequest"
            }
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/task/{user}/{taskId}": {
      "get": {
        "operationId": "Admin_TaskGet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTaskGetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/tasks/{user}": {
      "get": {
        "operationId": "Admin_TaskList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiTaskListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    }
  },
  "definitions": {
    "TaskListResponseTask": {
      "type": "object",
      "properties": {
        "taskId": {
          "type": "string",
          "format": "uint64"
        },
        "user": {
          "type": "string"
        },
        "task": {
          "type": "string"
        },
        "dueDate": {
          "type": "string"
        }
      }
    },
    "apiTaskCreateRequest": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string"
        },
        "task": {
          "type": "string"
        },
        "dueDate": {
          "type": "string"
        }
      }
    },
    "apiTaskCreateResponse": {
      "type": "object"
    },
    "apiTaskDeleteRequest": {
      "type": "object",
      "properties": {
        "taskId": {
          "type": "string",
          "format": "uint64"
        },
        "user": {
          "type": "string"
        }
      }
    },
    "apiTaskDeleteResponse": {
      "type": "object"
    },
    "apiTaskGetResponse": {
      "type": "object",
      "properties": {
        "taskId": {
          "type": "string",
          "format": "uint64"
        },
        "user": {
          "type": "string"
        },
        "task": {
          "type": "string"
        },
        "dueDate": {
          "type": "string"
        }
      }
    },
    "apiTaskListResponse": {
      "type": "object",
      "properties": {
        "Tasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TaskListResponseTask"
          }
        }
      }
    },
    "apiTaskUpdateRequest": {
      "type": "object",
      "properties": {
        "taskId": {
          "type": "string",
          "format": "uint64"
        },
        "user": {
          "type": "string"
        },
        "task": {
          "type": "string"
        },
        "dueDate": {
          "type": "string"
        }
      }
    },
    "apiTaskUpdateResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}